Had Adit diagnose Javascript issue where the page would not scroll after opening and closing any of the portfolio panels on mobile breakpoint. Here are Adit’s notes:
“the issue is this script: http://yougotkeith.github.io/keith/js/main.js
somehow it doesn't remove "overflow-hidden" class from body tag.
the issue code is on line 79 to 86 there is script:

$('body').removeClass('overflow-hidden')

on line 79: mq == 'mobile' should be return true and remove the
"overflow-hidden" class from body but, but some how doesn't work as
expected.

the easiest way to fix it you can put
$('body').removeClass('overflow-hidden') on line 78.”